adds edit routing
